Emerald: eliminate uint64 -> int64 conversions

+++ analyzer/emerald/emerald.go
@@ -283,6 +283,6 @@ func (m *Main) queueBlockAndTransactionInserts(batch *storage.QueryBatch, data *
 		blockData.Size,
 	)
 
-	emitRoundBatch(batch, int64(data.Round), blockData)
+	emitRoundBatch(batch, data.Round, blockData)
 	return nil
 }
